+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Stop shadowing Gradle plugin dependencies, manually repack only JPS Module
 - Load the `com.intellij` module by default when creating the IntelliJ Platform dependency
+- Load the `com.intellij` module with its optional dependencies for tests classpath as a cleaner fix for [IJPL-180516](https://youtrack.jetbrains.com/issue/IJPL-180516/Gradle-tests-fail-without-transitive-modules-jars-of-com.intellij-in-classpath)
 
 ### Fixed